+++ R/globals.R
@@ -1,4 +1,0 @@
-utils::globalVariables(c(
-  "hrrp_payment_periods",
-  "hrrp_performance_periods"
-))