+++ R/readmit-package.R
@@ -1,7 +1,6 @@
+#' readmit: R Interface to the Readmission Risk Pool
+#'
+#' `readmit` provides an R interface to the Readmission Risk Pool platform.
+#'
 #' @keywords internal
 "_PACKAGE"
-
-## usethis namespace: start
-#' @importFrom rlang .data
-## usethis namespace: end
-NULL