+++ tests/testthat/test-hrrp_get_dates.R
@@ -1,3 +1,0 @@
-test_that("No argument error", {
-  expect_error(hrrp_get_dates())
-})