+++ NEWS.md
@@ -1,5 +1,7 @@
 # readmit (development version)
 
+* Repurposing `readmit` as an R interface to the Readmission Risk Pool platform.
+
 # readmit 0.0.1
 
 * Initial CRAN submission.